new commit

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int i = 0;
-	long j = 1, k = 2;
+	long long j = 1, k = 2;
 
 	printf("%ld, %ld, ", j, k);
 	while (i < 98)
